% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsegICBM-package.R
\docType{package}
\name{ggsegICBM-package}
\alias{ggsegICBM}
\alias{ggsegICBM-package}
\title{ggsegICBM: ICBM Atlas for the 'ggseg' Ecosystem}
\description{
ICBM white matter atlas for the 'ggseg' ecosystem. Provides a unified 'ggseg_atlas' object with 3D mesh geometry, for use with 'ggseg3d'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ggseg/ggsegICBM}
  \item Report bugs at \url{https://github.com/ggseg/ggsegICBM/issues}
}

}
\author{
\strong{Maintainer}: Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID})

Authors:
\itemize{
  \item Didac Vidal-Piñeiro \email{d.v.pineiro@psykologi.uio.no} (\href{https://orcid.org/0000-0001-9997-9156}{ORCID})
}

}
\keyword{internal}
