% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icbm}
\alias{icbm}
\title{ICBM White Matter Atlas}
\format{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (subcortical).
}
\usage{
data(icbm)
}
\description{
Brain atlas for the ICBM DTI-81 white matter labels with 26 tracts.
Contains 3D mesh geometry for \code{\link[ggseg3d:ggseg3d]{ggseg3d::ggseg3d()}}. No 2D polygon
geometry is available for this atlas.
}
\examples{
data(icbm)
icbm
}
\references{
Mori et al., MRI Atlas of Human White Matter. Elsevier,
Amsterdam, The Netherlands (2005)
(\href{https://www.elsevier.com/books/mri-atlas-of-human-white-matter/mori/978-0-444-51741-8}{Elsevier})
}
\concept{ggseg_atlases}
\keyword{datasets}
